change code to work with different base meshes and added exception handling
data/hm08.config changed for the select groups, delete the double
definitions in extraproperties.py and operators/extractclothes.py
and changed function in selection to work dynamically

sanity checks for clothes now tests existence of groups on human when
called with two objects. Atm this is done only in case of make button
(without that system created empty mhclo file)

core_makeclothes_functionality.py contains json-loader for mesh.config, I had
to move it here. I had to divide __init__ function, because it cannot return
an error. I tried to create a few exceptions like a directory called
name.mhclo to see what happens. Now we got a message and not these horrible
exceptions in Blender. So I put "try" into write procedures as well and now
return sth. not None

there are some more exceptions we had to consider in future. For an
automatic detection of Mesh-config there must be an empty group called
_mesh_<something> in the basemesh object. If it is there, program looks for
<something>.config in data directory.  Otherwise always hm08.config is
used. The possibility not to take "head" for makeclothes itself has
still to be added. Markashuman will now also announce the meshtype.