added a functionality to calculate max-poles.
Without mentioning maxpoles makehuman
is at the moment limited to 16. The test function maybe changed to also
supply and array of vertex numbers later. direct calculation is made
in mhmesh. The fixes of Aranuvir I must include in mhmesh not to run
into trouble. Hopefully merging will not create conflicts