changes for marking a human
when one object was already marked as a human the user now can mark
another one to be a human and the first will be marked as clothes then.

When you ever tried to find an object which was accidentally marked as a
human you will never find it again. (except  people do know what
custom properties are)

I also added a code which avoids a division by zero error for the
determinante in case of rigid groups. For normal meshes we could use a
fourth "closest vertex" (which is not available for 3 vertices). But
there some additional lines must be changed.