fixes: deleting of vertices must be changed
For extract clothes delete function for vertices assumed each vertex in only one group,
everything worked fine, when it was the first group. The Helper was the
second group so all was deleted. And then we have a valid mesh of no
vertices :)

I also changed the text Extract clothes ... it is
"Optional base for clothes:"
+ "extract from helper" button now.

Since it is only one technique it makes sense this way.